@@ -1,7 +1,8 @@
 [package]
 name = "uniswap-sdk-core"
-version = "5.1.0"
-edition = "2021"
+version = "5.2.0"
+edition = "2024"
+rust-version = "1.85"
 authors = ["malik <aremumalik05@gmail.com>", "Shuhui Luo <twitter.com/aureliano_law>"]
 description = "The Uniswap SDK Core in Rust provides essential functionality for interacting with the Uniswap decentralized exchange"
 license = "MIT"
@@ -12,11 +13,11 @@ keywords = ["sdk-core", "ethereum", "sdk"]
 exclude = [".github", ".gitignore", "rustfmt.toml"]
 
 [dependencies]
-alloy-primitives = { version = "1.0", default-features = false, features = ["map-fxhash"] }
+alloy-primitives = { version = "1.3", default-features = false, features = ["map-fxhash"] }
 bnum = "0.12.0"
 derive_more = { version = "2", default-features = false, features = ["deref", "from"] }
 eth_checksum = { version = "0.1.2", optional = true }
-fastnum = { version = "0.2.3", default-features = false, features = ["numtraits"] }
+fastnum = { version = "0.2.10", default-features = false, features = ["numtraits"] }
 lazy_static = "1.5"
 num-integer = { version = "0.1", default-features = false }
 num-traits = { version = "0.2.19", default-features = false, features = ["libm"] }
